move exceptions to errors module
